Refactor according to review comments
