+++ tests/components/config/test_config_entries.py
@@ -12,13 +12,12 @@ from homeassistant.components.config import config_entries
 from homeassistant.config_entries import HANDLERS
 from homeassistant.core import callback
 from homeassistant.generated import config_flows
-from homeassistant.helpers import config_validation as cv, device_registry as dr
+from homeassistant.helpers import config_validation as cv
 from homeassistant.setup import async_setup_component
 
 from tests.common import (
     MockConfigEntry,
     MockModule,
-    mock_device_registry,
     mock_entity_platform,
     mock_integration,
 )
@@ -1034,259 +1033,3 @@ async def test_ignore_flow_nonexisting(hass, hass_ws_client):
 
     assert not response["success"]
     assert response["error"]["code"] == "not_found"
-
-
-async def test_remove_config_entry_from_device(hass, hass_ws_client):
-    """Test removing config entry from device."""
-    assert await async_setup_component(hass, "config", {})
-    ws_client = await hass_ws_client(hass)
-    device_registry = mock_device_registry(hass)
-
-    can_remove = False
-
-    async def async_remove_config_entry_device(hass, config_entry, device_entry):
-        return can_remove
-
-    mock_integration(
-        hass,
-        MockModule(
-            "comp1", async_remove_config_entry_device=async_remove_config_entry_device
-        ),
-    )
-    mock_integration(
-        hass,
-        MockModule(
-            "comp2", async_remove_config_entry_device=async_remove_config_entry_device
-        ),
-    )
-
-    entry_1 = MockConfigEntry(
-        domain="comp1",
-        title="Test 1",
-        source="bla",
-    )
-    entry_1.supports_remove_device = True
-    entry_1.add_to_hass(hass)
-
-    entry_2 = MockConfigEntry(
-        domain="comp1",
-        title="Test 1",
-        source="bla",
-    )
-    entry_2.supports_remove_device = True
-    entry_2.add_to_hass(hass)
-
-    device_registry.async_get_or_create(
-        config_entry_id=entry_1.entry_id,
-        connections={(dr.CONNECTION_NETWORK_MAC, "12:34:56:AB:CD:EF")},
-    )
-    device_entry = device_registry.async_get_or_create(
-        config_entry_id=entry_2.entry_id,
-        connections={(dr.CONNECTION_NETWORK_MAC, "12:34:56:AB:CD:EF")},
-    )
-    assert device_entry.config_entries == {entry_1.entry_id, entry_2.entry_id}
-
-    # Try removing a config entry from the device, it should fail because
-    # async_remove_config_entry_device returns False
-    await ws_client.send_json(
-        {
-            "id": 5,
-            "type": "config_entries/remove_device",
-            "entry_id": entry_1.entry_id,
-            "device_id": device_entry.id,
-        }
-    )
-    response = await ws_client.receive_json()
-
-    assert not response["success"]
-    assert response["error"]["code"] == "unknown_error"
-
-    # Make async_remove_config_entry_device return True
-    can_remove = True
-
-    # Remove the 1st config entry
-    await ws_client.send_json(
-        {
-            "id": 6,
-            "type": "config_entries/remove_device",
-            "entry_id": entry_1.entry_id,
-            "device_id": device_entry.id,
-        }
-    )
-    response = await ws_client.receive_json()
-
-    assert response["success"]
-    assert response["result"] is True
-
-    # Check that the config entry was removed from the device
-    assert device_registry.async_get(device_entry.id).config_entries == {
-        entry_2.entry_id
-    }
-
-    # Remove the 2nd config entry
-    await ws_client.send_json(
-        {
-            "id": 7,
-            "type": "config_entries/remove_device",
-            "entry_id": entry_2.entry_id,
-            "device_id": device_entry.id,
-        }
-    )
-    response = await ws_client.receive_json()
-
-    assert response["success"]
-    assert response["result"] is True
-
-    # This was the last config entry, the device is removed
-    assert not device_registry.async_get(device_entry.id)
-
-
-async def test_remove_config_entry_from_device_fails(hass, hass_ws_client):
-    """Test removing config entry from device failing cases."""
-    assert await async_setup_component(hass, "config", {})
-    ws_client = await hass_ws_client(hass)
-    device_registry = mock_device_registry(hass)
-
-    async def async_remove_config_entry_device(hass, config_entry, device_entry):
-        return True
-
-    mock_integration(
-        hass,
-        MockModule("comp1"),
-    )
-    mock_integration(
-        hass,
-        MockModule(
-            "comp2", async_remove_config_entry_device=async_remove_config_entry_device
-        ),
-    )
-
-    entry_1 = MockConfigEntry(
-        domain="comp1",
-        title="Test 1",
-        source="bla",
-    )
-    entry_1.add_to_hass(hass)
-
-    entry_2 = MockConfigEntry(
-        domain="comp2",
-        title="Test 1",
-        source="bla",
-    )
-    entry_2.supports_remove_device = True
-    entry_2.add_to_hass(hass)
-
-    entry_3 = MockConfigEntry(
-        domain="comp3",
-        title="Test 1",
-        source="bla",
-    )
-    entry_3.supports_remove_device = True
-    entry_3.add_to_hass(hass)
-
-    device_registry.async_get_or_create(
-        config_entry_id=entry_1.entry_id,
-        connections={(dr.CONNECTION_NETWORK_MAC, "12:34:56:AB:CD:EF")},
-    )
-    device_registry.async_get_or_create(
-        config_entry_id=entry_2.entry_id,
-        connections={(dr.CONNECTION_NETWORK_MAC, "12:34:56:AB:CD:EF")},
-    )
-    device_entry = device_registry.async_get_or_create(
-        config_entry_id=entry_3.entry_id,
-        connections={(dr.CONNECTION_NETWORK_MAC, "12:34:56:AB:CD:EF")},
-    )
-
-    fake_entry_id = "abc123"
-    assert entry_1.entry_id != fake_entry_id
-    fake_device_id = "abc123"
-    assert device_entry.id != fake_device_id
-
-    # Try removing a non existing config entry from the device
-    await ws_client.send_json(
-        {
-            "id": 5,
-            "type": "config_entries/remove_device",
-            "entry_id": fake_entry_id,
-            "device_id": device_entry.id,
-        }
-    )
-    response = await ws_client.receive_json()
-
-    assert not response["success"]
-    assert response["error"]["code"] == "not_found"
-    assert response["error"]["message"] == "Config entry not found"
-
-    # Try removing a config entry which does not support removal from the device
-    await ws_client.send_json(
-        {
-            "id": 6,
-            "type": "config_entries/remove_device",
-            "entry_id": entry_1.entry_id,
-            "device_id": device_entry.id,
-        }
-    )
-    response = await ws_client.receive_json()
-
-    assert not response["success"]
-    assert response["error"]["code"] == "home_assistant_error"
-    assert (
-        response["error"]["message"] == "Config entry does not support device removal"
-    )
-
-    # Try removing a config entry from a device which does not exist
-    await ws_client.send_json(
-        {
-            "id": 7,
-            "type": "config_entries/remove_device",
-            "entry_id": entry_2.entry_id,
-            "device_id": fake_device_id,
-        }
-    )
-    response = await ws_client.receive_json()
-
-    assert not response["success"]
-    assert response["error"]["code"] == "home_assistant_error"
-    assert response["error"]["message"] == "Unknown device"
-
-    # Try removing a config entry from a device which it's not connected to
-    await ws_client.send_json(
-        {
-            "id": 8,
-            "type": "config_entries/remove_device",
-            "entry_id": entry_2.entry_id,
-            "device_id": device_entry.id,
-        }
-    )
-    response = await ws_client.receive_json()
-
-    assert response["success"]
-
-    await ws_client.send_json(
-        {
-            "id": 9,
-            "type": "config_entries/remove_device",
-            "entry_id": entry_2.entry_id,
-            "device_id": device_entry.id,
-        }
-    )
-    response = await ws_client.receive_json()
-
-    assert not response["success"]
-    assert response["error"]["code"] == "home_assistant_error"
-    assert response["error"]["message"] == "Config entry not in device"
-
-    # Try removing a config entry which can't be loaded from a device
-    await ws_client.send_json(
-        {
-            "id": 10,
-            "type": "config_entries/remove_device",
-            "entry_id": entry_3.entry_id,
-            "device_id": device_entry.id,
-        }
-    )
-    response = await ws_client.receive_json()
-
-    assert not response["success"]
-    assert response["error"]["code"] == "home_assistant_error"
-    assert response["error"]["message"] == "Integration not found"
