+++ tests/components/config/test_device_registry.py
@@ -3,8 +3,14 @@ import pytest
 
 from homeassistant.components.config import device_registry
 from homeassistant.helpers import device_registry as helpers_dr
+from homeassistant.setup import async_setup_component
 
-from tests.common import mock_device_registry
+from tests.common import (
+    MockConfigEntry,
+    MockModule,
+    mock_device_registry,
+    mock_integration,
+)
 from tests.components.blueprint.conftest import stub_blueprint_populate  # noqa: F401
 
 
@@ -126,3 +132,239 @@ async def test_update_device(hass, client, registry, payload_key, payload_value)
     assert getattr(device, payload_key) == payload_value
 
     assert isinstance(device.disabled_by, (helpers_dr.DeviceEntryDisabler, type(None)))
+
+
+async def test_remove_config_entry_from_device(hass, hass_ws_client):
+    """Test removing config entry from device."""
+    assert await async_setup_component(hass, "config", {})
+    ws_client = await hass_ws_client(hass)
+    device_registry = mock_device_registry(hass)
+
+    mock_integration(
+        hass,
+        MockModule("comp1", support_remove_from_device=True),
+    )
+    mock_integration(
+        hass,
+        MockModule("comp2", support_remove_from_device=True),
+    )
+
+    entry_1 = MockConfigEntry(
+        domain="comp1",
+        title="Test 1",
+        source="bla",
+    )
+    entry_1.supports_remove_device = True
+    entry_1.add_to_hass(hass)
+
+    entry_2 = MockConfigEntry(
+        domain="comp1",
+        title="Test 1",
+        source="bla",
+    )
+    entry_2.supports_remove_device = True
+    entry_2.add_to_hass(hass)
+
+    device_registry.async_get_or_create(
+        config_entry_id=entry_1.entry_id,
+        connections={(helpers_dr.CONNECTION_NETWORK_MAC, "12:34:56:AB:CD:EF")},
+    )
+    device_entry = device_registry.async_get_or_create(
+        config_entry_id=entry_2.entry_id,
+        connections={(helpers_dr.CONNECTION_NETWORK_MAC, "12:34:56:AB:CD:EF")},
+    )
+    assert device_entry.config_entries == {entry_1.entry_id, entry_2.entry_id}
+
+    # Remove the 1st config entry
+    await ws_client.send_json(
+        {
+            "id": 5,
+            "type": "config/device_registry/remove_config_entry",
+            "config_entry_id": entry_1.entry_id,
+            "device_id": device_entry.id,
+        }
+    )
+    response = await ws_client.receive_json()
+
+    assert response["success"]
+    assert response["result"]["config_entries"] == [entry_2.entry_id]
+
+    # Check that the config entry was removed from the device
+    assert device_registry.async_get(device_entry.id).config_entries == {
+        entry_2.entry_id
+    }
+
+    # Remove the 2nd config entry
+    await ws_client.send_json(
+        {
+            "id": 6,
+            "type": "config/device_registry/remove_config_entry",
+            "config_entry_id": entry_2.entry_id,
+            "device_id": device_entry.id,
+        }
+    )
+    response = await ws_client.receive_json()
+
+    assert response["success"]
+    assert response["result"] is None
+
+    # This was the last config entry, the device is removed
+    assert not device_registry.async_get(device_entry.id)
+
+
+async def test_remove_config_entry_from_device_fails(hass, hass_ws_client):
+    """Test removing config entry from device failing cases."""
+    assert await async_setup_component(hass, "config", {})
+    ws_client = await hass_ws_client(hass)
+    device_registry = mock_device_registry(hass)
+
+    mock_integration(
+        hass,
+        MockModule("comp1"),
+    )
+    mock_integration(
+        hass,
+        MockModule("comp2", support_remove_from_device=True),
+    )
+
+    entry_1 = MockConfigEntry(
+        domain="comp1",
+        title="Test 1",
+        source="bla",
+    )
+    entry_1.add_to_hass(hass)
+
+    entry_2 = MockConfigEntry(
+        domain="comp2",
+        title="Test 1",
+        source="bla",
+    )
+    entry_2.supports_remove_device = True
+    entry_2.add_to_hass(hass)
+
+    entry_3 = MockConfigEntry(
+        domain="comp3",
+        title="Test 1",
+        source="bla",
+    )
+    entry_3.supports_remove_device = True
+    entry_3.add_to_hass(hass)
+
+    device_registry.async_get_or_create(
+        config_entry_id=entry_1.entry_id,
+        connections={(helpers_dr.CONNECTION_NETWORK_MAC, "12:34:56:AB:CD:EF")},
+    )
+    device_registry.async_get_or_create(
+        config_entry_id=entry_2.entry_id,
+        connections={(helpers_dr.CONNECTION_NETWORK_MAC, "12:34:56:AB:CD:EF")},
+    )
+    device_entry = device_registry.async_get_or_create(
+        config_entry_id=entry_3.entry_id,
+        connections={(helpers_dr.CONNECTION_NETWORK_MAC, "12:34:56:AB:CD:EF")},
+    )
+    assert device_entry.config_entries == {
+        entry_1.entry_id,
+        entry_2.entry_id,
+        entry_3.entry_id,
+    }
+
+    fake_entry_id = "abc123"
+    assert entry_1.entry_id != fake_entry_id
+    fake_device_id = "abc123"
+    assert device_entry.id != fake_device_id
+
+    # Try removing a non existing config entry from the device - does nothing
+    await ws_client.send_json(
+        {
+            "id": 5,
+            "type": "config/device_registry/remove_config_entry",
+            "config_entry_id": fake_entry_id,
+            "device_id": device_entry.id,
+        }
+    )
+    response = await ws_client.receive_json()
+    assert response["success"]
+    assert set(response["result"]["config_entries"]) == {
+        entry_1.entry_id,
+        entry_2.entry_id,
+        entry_3.entry_id,
+    }
+
+    # Try removing a config entry which does not support removal from the device
+    await ws_client.send_json(
+        {
+            "id": 6,
+            "type": "config/device_registry/remove_config_entry",
+            "config_entry_id": entry_1.entry_id,
+            "device_id": device_entry.id,
+        }
+    )
+    response = await ws_client.receive_json()
+
+    assert not response["success"]
+    assert response["error"]["code"] == "home_assistant_error"
+    assert (
+        response["error"]["message"] == "Config entry does not support device removal"
+    )
+
+    # Try removing a config entry from a device which does not exist
+    await ws_client.send_json(
+        {
+            "id": 7,
+            "type": "config/device_registry/remove_config_entry",
+            "config_entry_id": entry_2.entry_id,
+            "device_id": fake_device_id,
+        }
+    )
+    response = await ws_client.receive_json()
+
+    assert not response["success"]
+    assert response["error"]["code"] == "home_assistant_error"
+    assert response["error"]["message"] == "Unknown device"
+
+    # Try removing a config entry from a device which it's not connected to - does nothing
+    await ws_client.send_json(
+        {
+            "id": 8,
+            "type": "config/device_registry/remove_config_entry",
+            "config_entry_id": entry_2.entry_id,
+            "device_id": device_entry.id,
+        }
+    )
+    response = await ws_client.receive_json()
+
+    assert response["success"]
+    assert set(response["result"]["config_entries"]) == {
+        entry_1.entry_id,
+        entry_3.entry_id,
+    }
+
+    await ws_client.send_json(
+        {
+            "id": 9,
+            "type": "config/device_registry/remove_config_entry",
+            "config_entry_id": entry_2.entry_id,
+            "device_id": device_entry.id,
+        }
+    )
+    response = await ws_client.receive_json()
+
+    assert response["success"]
+    assert set(response["result"]["config_entries"]) == {
+        entry_1.entry_id,
+        entry_3.entry_id,
+    }
+
+    # Try removing a config entry which can't be loaded from a device - allowed
+    await ws_client.send_json(
+        {
+            "id": 10,
+            "type": "config/device_registry/remove_config_entry",
+            "config_entry_id": entry_3.entry_id,
+            "device_id": device_entry.id,
+        }
+    )
+    response = await ws_client.receive_json()
+
+    assert response["success"]
+    assert set(response["result"]["config_entries"]) == {entry_1.entry_id}
